requirements.txt
playwright==1.54.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
trafilatura==2.0.0
fake-useragent==1.5.1
httpx==0.27.0
tqdm==4.66.4
sqlalchemy==2.0.31
pydantic==2.8.2
python-dotenv==1.0.1
toml==0.10.2
apscheduler==3.10.4
python-pptx==0.6.23
nltk==3.9.1
scikit-learn==1.5.1
umap-learn==0.5.6
bertopic==0.16.3
transformers==4.52.4
torch>=2.7.1
requests==2.32.3
feedparser==6.0.11